+++ packages/runtime/CHANGELOG.md
@@ -1,3 +1,12 @@
+## @lunora/runtime [1.0.0-alpha.103](https://github.com/anolilab/lunora/compare/@lunora/runtime@1.0.0-alpha.102...@lunora/runtime@1.0.0-alpha.103) (2026-09-07)
+
+
+### Dependencies
+
+* **@lunora/observability:** upgraded to 1.0.0-alpha.61
+* **@lunora/do:** upgraded to 1.0.0-alpha.123
+* **@lunora/shard-engine:** upgraded to 1.0.0-alpha.61
+
 ## @lunora/runtime [1.0.0-alpha.102](https://github.com/anolilab/lunora/compare/@lunora/runtime@1.0.0-alpha.101...@lunora/runtime@1.0.0-alpha.102) (2026-09-07)
 
 ### Bug Fixes