+++ packages/vue/CHANGELOG.md
@@ -1,3 +1,10 @@
+## @lunora/vue [1.0.0-alpha.124](https://github.com/anolilab/lunora/compare/@lunora/vue@1.0.0-alpha.123...@lunora/vue@1.0.0-alpha.124) (2026-09-07)
+
+
+### Dependencies
+
+* **@lunora/client:** upgraded to 1.0.0-alpha.88
+
 ## @lunora/vue [1.0.0-alpha.123](https://github.com/anolilab/lunora/compare/@lunora/vue@1.0.0-alpha.122...@lunora/vue@1.0.0-alpha.123) (2026-09-07)